// backend/src/types/code/vehicleCode.ts
export enum VehicleErrorCode {
  PLATE_ALREADY_EXISTS = "PLATE_ALREADY_EXISTS",
  VEHICLE_CREATE_FAILED = "VEHICLE_CREATE_FAILED",
  VEHICLE_NOT_FOUND = "VEHICLE_NOT_FOUND",
  VEHICLE_FETCH_FAILED = "VEHICLE_FETCH_FAILED",
  VEHICLE_DELETE_FAILED = "VEHICLE_DELETE_FAILED",
  VEHICLE_UPDATE_FAILED = "VEHICLE_UPDATE_FAILED",
}